feat: move connection components to caldera package
